changing script...

+++ task1/run.sh
@@ -30,9 +30,18 @@ mkdir "$backup_folder";
 for file in $(find "$input_folder" -name "*.$extension"); do
 
 	name=$(echo $file | rev | cut -d'/' -f-1 | rev);
-	cp $file "$backup_folder/$name";
+
+	if [[ -f "$backup_folder/$name" ]]; then
+		counter=1;
+		while [[ -f "$backup_folder/$name.$counter" ]]; do
+			(( counter += 1 ));
+		done;
+
+		cp $file "$backup_folder/$name.$counter";
+	else 
+		cp $file "$backup_folder/$name";
+	fi;
 done;
 
 tar -czf "$backup_archive" "$backup_folder"
 
-echo "done"
